change: PokerTableに記載の処理を再整理する
